examples/gradio_app/requirements.txt
gradio
torch
Pillow
opencv-python
numpy